proto/bot/webhook-python/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-telegram-bot==22.1
python-dotenv==1.1.0

